Added ddb_json_parser
