+++ pyproject.toml
@@ -8,6 +8,7 @@ requires = ["maturin>=0.13,<0.14"]
 
 [tool.maturin]
 sdist-include = ["build.rs", "Cargo.lock", "include/**/*"]
+python-source = ["python"]
 strip = true
 
 [tool.black]
