+++ python/orjson_ddb/__init__.py
@@ -1,0 +1,4 @@
+from .orjson_ddb import loads
+from .boto_integration import ddb_json_parser
+
+
