+++ test/ddb/test_boto_integration.py
@@ -4,24 +4,7 @@ from decimal import Decimal
 import pytest
 from moto import mock_dynamodb2
 import boto3
-from botocore.parsers import PROTOCOL_PARSERS
-import orjson_ddb
-
-
-@contextlib.contextmanager
-def json_ddb_parser():
-    # Start the timer
-    _parser = PROTOCOL_PARSERS.get("json")
-    old_parser = getattr(_parser, "_handle_json_body")
-
-    def new_parser(self, raw_body, shape):
-        return orjson_ddb.loads(raw_body)
-
-    setattr(_parser, "_handle_json_body", new_parser)
-
-    yield
-
-    setattr(_parser, "_handle_json_body", old_parser)
+from orjson_ddb import ddb_json_parser
 
 
 TABLE_NAME = "some_table"
@@ -71,7 +54,7 @@ def put_item_in_ddb(dynamodb_mocked):
 
 def test_ddb_boto3_get_item(dynamodb_mocked, dynamo_table, put_item_in_ddb):
     dynamodb_client = boto3.client("dynamodb", region_name="us-east-1")
-    with json_ddb_parser():
+    with ddb_json_parser():
         resp = dynamodb_client.get_item(
             TableName=TABLE_NAME,
             Key={"pk": {"S": "pk1"}, "sk": {"S": "sk1"}}
@@ -82,7 +65,7 @@ def test_ddb_boto3_get_item(dynamodb_mocked, dynamo_table, put_item_in_ddb):
 
 def test_ddb_boto3_query_items(dynamodb_mocked, dynamo_table, put_item_in_ddb):
     dynamodb_client = boto3.client("dynamodb", region_name="us-east-1")
-    with json_ddb_parser():
+    with ddb_json_parser():
         resp = dynamodb_client.query(
             TableName=TABLE_NAME,
             KeyConditionExpression='pk = :pk',
